+++ src/settings/Validation/notice-policy/notice-policy.js
@@ -2,6 +2,7 @@ import FormValidator from '../engine/FormValidator';
 import general from './general';
 import loanNotices from './loan-notices';
 import requestNotices from './request-notices';
+import feeFineNotices from './fee-fine-notices';
 import { NoticePolicy } from '../../Models/NoticePolicy';
 
 export default function (policy) {
@@ -10,6 +11,7 @@ export default function (policy) {
     ...general(),
     ...loanNotices(noticePolicy),
     ...requestNotices(noticePolicy),
+    ...feeFineNotices(noticePolicy),
   };
   const formValidator = new FormValidator(config);